Move window.py into game folder
